some changes, ill clean this up in a hot sec
